api/requirements.txt
fastapi==0.70.0
gunicorn==20.1.0
uvicorn==0.15.0
requests==2.26.0
mysql-connector-python==8.0.26
requests==2.26.0
fastapi-pagination==0.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
pyexcel==0.6.7
pyexcel-xls==0.7.0
